l10n_ES_account_report: Correcciones y mejoras (antiguo módulo account_ES_reporting), l10n_chart_ES: Eliminación informes vacios, c2c_finance_report: Pequeñas mejoras y traducciones
