--- l10n_es/account_ES_reporting/__terp__.py
@@ -1,21 +1,0 @@
-{
-    "name" : "Reporting of Balancesheet for accounting Spain",
-    "version" : "1.0",
-    "depends" : ["account","l10n_chart_ES"],
-    "author" : "Tiny and Acysos S.L.",
-    "description": """Financial and accounting reporting Spain""",
-    "category" : "Generic Modules/Accounting",
-    "init_xml" : [ ],
-    "demo_xml" : [ ],
-    "update_xml" : [
-                    "account_view.xml",
-                    "account_report.xml",
-                    "account_data.xml",
-                    "account_report_ES.xml",
-                    ],
-#    "translations" : {
-#        "fr": "i18n/french_fr.csv"
-#    },
-    "active": False,
-    "installable": True
-}
