--- l10n_es/account_ES_reporting/account_report.xml
@@ -1,18 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<terp>
-    <data>
-<!--        <report id="balancesheet_report"
-            string="BalanceSheet Report"
-            model="account.report.bs"
-            name="account.report.bs"
-            rml="addons/account_report_bs/report/account_report_bs.rml"
-            auto="False"/>-->
-        <wizard
-            string="Account balance"
-            model="account.report.bs"
-            name="account.account.balancesheet.report"
-            keyword="client_print_multi"
-            id="wizard_balance_report"/>
-
-    </data>
-</terp>
