--- l10n_es/account_ES_reporting/account_view.xml
@@ -1,64 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<terp>
-<data>
-
-    <record model="ir.ui.view" id="view_account_report_bs_form">
-        <field name="name">account.report.bs.form</field>
-        <field name="model">account.report.bs</field>
-        <field name="type">form</field>
-        <field name="arch" type="xml">
-            <form string="Account reporting">
-                <notebook>
-                <page string="General">
-                    <field name="name" select="1" colspan="3"/>
-                    <field name="code" select="1"/>
-                    <field name="parent_id" select = "1" />
-                    <field name="sequence"/>
-                    <field name="color_font"/>
-                    <field name="color_back"/>
-                    <field name="font_style" />
-                    <field name="report_type" />
-                    <newline/>
-                    <separator string="Accounts" colspan="4"/>
-                        <field name="account_id" colspan="4" nolable="1" domain="[('type','=','view')]"/>
-                </page><page string="Notes">
-                    <field name="note" nolabel="1" colspan="4"/>
-                </page>
-                </notebook>
-            </form>
-        </field>
-    </record>
-
-        <record model="ir.ui.view" id="view_account_report_tree_bs">
-            <field name="name">account.report.report.tree.bs</field>
-            <field name="model">account.report.bs</field>
-            <field name="type">tree</field>
-            <field name="field_parent">child_id</field>
-            <field name="arch" type="xml">
-                <tree string="Account reporting">
-                <field name="code"/>
-                <field name="name"/>
-            </tree>
-            </field>
-        </record>
-        <record model="ir.actions.act_window" id="action_account_tree_bs">
-            <field name="type">ir.actions.act_window</field>
-            <field name="res_model">account.report.bs</field>
-            <field name="domain">[('parent_id','=',False)]</field>
-            <field name="view_type">tree</field>
-            <field name="view_id" ref="view_account_report_tree_bs"/>
-        </record>
-        <menuitem name="Financial Management/Configuration/Balance Sheet Report/Balance Sheet Report" id="action_account_report_bs_form" action="action_account_tree_bs"/>
-
-        <record model="ir.actions.act_window" id="acc_bs_report_action_form">
-            <field name="type">ir.actions.act_window</field>
-            <field name="res_model">account.report.bs</field>
-            <field name="view_type">form</field>
-            <field name="view_id" ref="view_account_report_bs_form"/>
-        </record>
-        <menuitem name="Financial Management/Configuration/Balance Sheet Report/Balance Sheet Rrport Form" id="bs_report_action_form" action="acc_bs_report_action_form"/>
-
-
-
-</data>
-</terp>
