+++ l10n_es/c2c_finance_report/account_balance.rml
@@ -134,7 +134,7 @@
     <para style="Standard">
       <font color="white"> </font>
     </para>
-    <blockTable colWidths="308.0,77.0,82.0,74.0" style="Table4">
+    <!--<blockTable colWidths="308.0,77.0,82.0,74.0" style="Table4">
       <tr>
         <td>
           <para style="P6">Total:</para>
@@ -149,7 +149,7 @@
           <para style="P7">[[ comma_me(sum_balance()) ]]</para>
         </td>
       </tr>
-    </blockTable>
+    </blockTable>-->
   </story>
 </document>
