+++ l10n_es/l10n_chart_ES/report/__init__.py
@@ -1,8 +1,1 @@
-import balance_oficial
-import balance_oficial_abreviado
-import pyg_oficial
-import pyg_oficial_abreviado
-import eigr_oficial
-import eigr_oficial_abreviado
-import efe_oficial
-import patrimonio_neto_oficial
+
